python-processor/requirements.txt
# Document Processing
pypdf2==3.0.1
langchain==0.1.0
langchain-community==0.0.10

# Embeddings - Updated versions for compatibility
sentence-transformers==3.0.1
transformers==4.44.0
huggingface-hub==0.24.0

# Azure
azure-search-documents==11.4.0
azure-core==1.29.5

# UI
gradio==4.12.0

# Utilities
python-dotenv==1.0.0
numpy==1.24.3
torch>=2.0.0
